+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/kid/controller/KidController.java
@@ -5,17 +5,12 @@ import com.kidsworld.kidsping.domain.kid.dto.request.KidMbtiDiagnosisRequest;
 import com.kidsworld.kidsping.domain.kid.dto.request.UpdateKidRequest;
 import com.kidsworld.kidsping.domain.kid.dto.response.*;
 import com.kidsworld.kidsping.domain.kid.service.KidService;
-import com.kidsworld.kidsping.domain.user.entity.User;
-import com.kidsworld.kidsping.domain.user.exception.UserNotFoundException;
-import com.kidsworld.kidsping.domain.user.service.UserService;
 import com.kidsworld.kidsping.global.common.dto.ApiResponse;
 import com.kidsworld.kidsping.global.exception.ExceptionCode;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -32,7 +27,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class KidController {
 
     private final KidService kidService;
-    private final UserService userService;
 
     /*
     자녀 프로필 생성
@@ -97,17 +91,4 @@ public class KidController {
     }
 
 
-    /*
-    회원 자녀 리스트 조회
-     */
-    @GetMapping("/list")
-    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
-    public ResponseEntity<ApiResponse<List<GetKidListResponse>>> getKidList(@AuthenticationPrincipal UserDetails userDetails) {
-        User user = userService.findByEmail(userDetails.getUsername())
-                .orElseThrow(() -> new UserNotFoundException());
-
-        List<GetKidListResponse> response = kidService.getKidsList(user.getId());
-        return ApiResponse.ok(ExceptionCode.OK.getCode(), response, "자녀 목록을 성공적으로 조회했습니다."
-        );
-    }
 }
