+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/kid/repository/KidRepository.java
@@ -2,7 +2,6 @@ package com.kidsworld.kidsping.domain.kid.repository;
 
 import com.kidsworld.kidsping.domain.kid.entity.Kid;
 
-import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -16,8 +15,4 @@ public interface KidRepository extends JpaRepository<Kid, Long> {
 
     @Query("select k from Kid k where k.id = :kidId and k.isDeleted = false")
     Optional<Kid> findKidBy(@Param("kidId") Long kidId);
-
-    //User의 kid
-    @Query("select k from Kid k where k.user.id = :userId and k.isDeleted = false")
-    List<Kid> findByUserIdAndIsDeletedFalse(@Param("userId") Long userId);
 }
