+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/kid/service/KidService.java
@@ -19,7 +19,6 @@ public interface KidService {
 
     List<GetKidMbtiHistoryResponse> getKidMbtiHistory(Long kidId);
 
-    List<GetKidListResponse> getKidsList(Long userId);
 
     void diagnoseKidMbti(KidMbtiDiagnosisRequest diagnosisRequest);
 }
