+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/user/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.kidsworld.kidsping.domain.user.controller;
 
+import com.kidsworld.kidsping.domain.user.dto.response.GetKidListResponse;
 import com.kidsworld.kidsping.domain.user.dto.request.LoginRequest;
 import com.kidsworld.kidsping.domain.user.dto.request.RegisterRequest;
 import com.kidsworld.kidsping.domain.user.dto.response.GetUserResponse;
@@ -14,12 +15,15 @@ import com.kidsworld.kidsping.global.jwt.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 
 @Slf4j
@@ -33,6 +37,9 @@ public class UserController {
     private final AuthenticationManager authenticationManager;
     private final JwtUtil jwtUtil;
 
+    /*
+    회원가입
+    */
     @PostMapping("/register")
     public ResponseEntity<?> registerUser(@RequestBody RegisterRequest registerRequest) {
         registerRequest.setPassword(passwordEncoder.encode(registerRequest.getPassword()));
@@ -40,6 +47,10 @@ public class UserController {
         return ResponseEntity.ok(RegisterResponse.builder().id(user.getId()).email(user.getEmail()).role(user.getRole()).build());
     }
 
+
+    /*
+    로그인
+    */
     @PostMapping("/login")
     public ResponseEntity<?> createAuthenticationToken(@RequestBody LoginRequest loginRequest) throws Exception {
         authenticationManager.authenticate(
@@ -53,6 +64,9 @@ public class UserController {
     }
 
 
+    /*
+    로그아웃
+    */
     @PostMapping("/logout")
     public ResponseEntity<ApiResponse<Void>> logout(@AuthenticationPrincipal UserDetails userDetails) {
 
@@ -63,6 +77,10 @@ public class UserController {
     }
 
 
+
+    /*
+    회원정보조회
+    */
     @GetMapping("/me")
     public ResponseEntity<ApiResponse<GetUserResponse>> getCurrentUser(@AuthenticationPrincipal UserDetails userDetails) {
         if (userDetails == null) {
@@ -85,5 +103,25 @@ public class UserController {
 
         return ApiResponse.ok(ExceptionCode.OK.getCode(), response, ExceptionCode.OK.getMessage());
     }
+
+
+    /*
+    회원 자녀 리스트 조회
+     */
+    @GetMapping("/{userId}/kidslist")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
+    public ResponseEntity<ApiResponse<List<GetKidListResponse>>> getKidList(@PathVariable("userId") Long userId,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        User user = userService.findByEmail(userDetails.getUsername())
+                .orElseThrow(UserNotFoundException::new);
+
+
+        List<GetKidListResponse> response = userService.getKidsList(user.getId());
+        return ApiResponse.ok(ExceptionCode.OK.getCode(), response, "자녀 목록을 성공적으로 조회했습니다.");
+    }
+
+
+
 
 }
