+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/user/dto/request/RegisterRequest.java
@@ -5,12 +5,14 @@ import com.kidsworld.kidsping.domain.user.entity.enums.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 
 @Data
 @Builder
 @AllArgsConstructor
+@NoArgsConstructor
 public class RegisterRequest implements Serializable {
     private String email;
     private String password;
