+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/user/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.kidsworld.kidsping.domain.user.service;
 
+import com.kidsworld.kidsping.domain.user.dto.response.GetKidListResponse;
 import com.kidsworld.kidsping.domain.user.dto.request.RegisterRequest;
 import com.kidsworld.kidsping.domain.user.entity.User;
 import com.kidsworld.kidsping.domain.user.repository.UserRepository;
@@ -7,9 +8,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -27,9 +31,19 @@ public class UserServiceImpl implements UserService {
         return userRepository.findByEmail(email);
     }
 
+
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         User user = userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException("User not found"));
         return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), new ArrayList<>());
+    }
+
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<GetKidListResponse> getKidsList(Long userId) {
+        return userRepository.findByUserIdAndIsDeletedFalse(userId).stream()
+                .map(GetKidListResponse::from)
+                .collect(Collectors.toList());
     }
 }
